+++ lib/src/config/app_data.dart
@@ -1,5 +1,5 @@
 import 'package:appquitanda/src/models/Iten_model.dart';
-import 'package:appquitanda/src/models/user_modelo.dart';
+import 'package:appquitanda/src/models/user_model.dart';
 import 'package:appquitanda/src/models/cart_item_model.dart';
 import 'package:appquitanda/src/models/order_model.dart';
 
@@ -58,7 +58,7 @@ List<CartItemModel> cartItems = [
 UserModel user = UserModel(
     cpf: '999.999.999-00',
     password: 'Sen888888ha',
-    phonenumber: '+12 345 567890123',
+    phone: '+12 345 567890123',
     email: 'emailcomum@indiferente.igual',
     name: 'nulidade');
