+++ pubspec.yaml
@@ -24,13 +24,17 @@ dependencies:
   shimmer: ^2.0.0
   get: ^4.6.5
   dio: ^4.0.6
-  
+  json_annotation: ^4.7.0  
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 
   flutter_launcher_icons: ^0.10.0
+
+  json_serializable: ^6.5.4
+
+  build_runner: ^2.1.10
 
 
   flutter_lints: ^2.0.0
